header useffect remove
