+++ src/components/HeaderContent/types.ts
@@ -1,0 +1,3 @@
+export type THeaderContent = {
+  phoneNumber: string;
+};
